+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 
 - Pass static string in the AntithesisErrorReporter.
+- Use static string map to store method signatures in instrumenters instead of relying on reflection.
 
 ### Deprecated